Fixed overflow in Rand vec base

+++ src/vec_base.cpp
@@ -178,10 +178,10 @@ void FullTrace::generate_vec_re( r_type rand_vec[], int r){
 
 void projected_FullTrace::generate_vec_im( std::complex<r_type> rand_vec[], int r){
 
-  int SUBDIM = this->parameters().SUBDIM_;
+  size_t SUBDIM = this->parameters().SUBDIM_;
 
 
-  for( int j = 0; j <  SUBDIM; j++ ){
+  for( size_t j = 0; j <  SUBDIM; j++ ){
     rand_vec [j] = 0;
     if( ( j - ( r - 1 ) ) % stride_ == 0){
       double phase = 2.0 * M_PI * this->rng().get() ;
@@ -194,10 +194,10 @@ void projected_FullTrace::generate_vec_im( std::complex<r_type> rand_vec[], int 
 
 void projected_FullTrace::generate_vec_re( r_type rand_vec[], int r){
 
-  int SUBDIM = this->parameters().SUBDIM_;
+  size_t SUBDIM = this->parameters().SUBDIM_;
 
   
-  for( int j = 0; j <  SUBDIM; j++ )
+  for( size_t j = 0; j <  SUBDIM; j++ )
     if( ( j - ( r - 1 ) ) % stride_ == 0)
       rand_vec [j] = (2.0*this->rng().get() - 1.)*sqrt(3.0)/ sqrt(double(SUBDIM/stride_));
